+++ packages/headless/src/features/facets/range-facets/numeric-facet-set/numeric-facet-set-slice.test.ts
@@ -57,6 +57,7 @@ describe('numeric-facet-set slice', () => {
       preventAutoSelect: false,
       sortCriteria: 'ascending',
       type: 'numericalRange',
+      rangeAlgorithm: 'even',
     });
   });